[FIX] PHAssetImageView 객체 하나로 통일(#11)
